Adding servers back from the LABRAD repository
Changed lakeshore218  back to what it was.  @nmGit, I think this will mess you up a bit, but not much.  All the functionality you wrote was already there, it was just stuck in a different repo for some reason.  I had to change it back or the ADR controls break.